Applied code review comments, attempt to fix clang build failure with type in loop.

+++ examples/energy-management-app/linux/main.cpp
@@ -130,7 +130,7 @@ static bool EnergyAppOptionHandler(const char * aProgram, chip::ArgParser::Optio
     {
     case kOptionApplication:
         spApp = nullptr;
-        for (uint16_t idx = 0; idx < sizeof(kValidApps) / sizeof(kValidApps[0]); idx++)
+        for (size_t idx = 0; idx < (sizeof(kValidApps) / sizeof(kValidApps[0])); idx++)
         {
             if (strcmp(kValidApps[idx], aValue) == 0)
             {
